Widen the remaining 32-bit pointer stashes to uintptr_t
Everything here is the same LP64 bug: decomp code that stores a pointer
in a u32/s32 field, or reconstructs one from four bytes, because on
GameCube a pointer was exactly a word. On x86-64 and arm64 the top half
is discarded and the value dereferences into nothing.

- HSD_IDTable keys and the objalloc heap cursors are now uintptr_t; the
  ID hash takes uintptr_t and narrows only its own result.
- The sislib text stream pushed and popped a cursor pointer one byte at
  a time. It now memcpy's sizeof(void*) and reserves that much.
- THPVideoDecode's return value is a pointer disguised as s32, so it and
  its callers (THPDec_80331340/803313D0, lbMthp, THPDecComp::unk_98)
  move to uintptr_t. __THPPrepBitStream masks with ~(uintptr_t)3 instead
  of 0xFFFFFFFC.
- hsd_3B33's bounds checks compared (u32)ptr + length; they now compare
  pointers directly.
- Stage code stored GObj/JObj pointers in u32 GroundVars slots. Those
  fields become typed pointers, which also removes the casts at every
  use and, in grcastle/grzebes, a couple of wrong union members that the
  casts were hiding.

Two fields turned out not to be pointers at all:

- itChicoritaLeafAttr::fighter is a duration, so it_80272818 returns s32
  and ftColl_8007B7FC (which takes an int) stops truncating a pointer.
- ftCo_0A01 passed `(Fighter_GObj*) blocked` to mpCheckFloor where
  `blocked` is the integer 0 it had just assigned. That is NULL spelled
  the long way.

__THPHuffGenerateSizeTable also takes u8* rather than int, dropping a
pointer-through-int round trip.

+++ src/melee/ft/ftcoll.c
@@ -1116,7 +1116,7 @@ bool ftColl_80077C60(Item* item, HitCapsule* hit, Fighter* fp,
             case It_Kind_Star:
                 hit->state = HitCapsule_Disabled;
                 item->xC34_damageDealt = 1;
-                ftColl_8007B7FC(fp, (int) it_80272818(item));
+                ftColl_8007B7FC(fp, it_80272818(item));
                 ft_PlaySFX(fp, 0xF9, 0x7F, 0x40);
                 ftCommon_8007EBAC(fp, 0x11, 0);
                 break;
