+++ src/sysdolphin/baselib/jobj.c
@@ -665,8 +665,8 @@ s32 JObjLoad(HSD_JObj* jobj, HSD_Joint* joint, HSD_JObj* parent)
             }
         }
     }
-    HSD_IDInsertToTable(NULL, (u32) joint, jobj);
-    jobj->id = (u32) joint;
+    HSD_IDInsertToTable(NULL, (uintptr_t) joint, jobj);
+    jobj->id = (uintptr_t) joint;
     return 0;
 }
 
@@ -1522,7 +1522,7 @@ void JObjRelease(HSD_Class* o)
     HSD_JOBJ_METHOD(jobj)->release_child(jobj);
 
     if (HSD_IDGetDataFromTable(NULL, jobj->id, NULL) == jobj) {
-        u32 id = jobj->id;
+        uintptr_t id = jobj->id;
         HSD_IDRemoveByIDFromTable(NULL, id);
     }
     if (jobj->scl != NULL) {
